Average calculation error due to ROUND() contains two versions one single argument and another with two arguments. We will use the two arguents to specify decimal places and the first argument is required to be numeric. Must include ROUNT(AVG(gpa)):: numeric,2) instead of ROUND(AVG(gpa), 2) AS avg_gpa,

+++ module_3/query_data.py
@@ -122,10 +122,10 @@ Question 3: What is the average GPA, GRE, GRE V, GRE AW of applicants who provid
 print("\n Question 3. Average GPA, GRE, GRE V, and GRE AW (excluding missing values):")
 q3 = """
 SELECT
-    ROUND(AVG(gpa), 2) AS avg_gpa,
-    ROUND(AVG(gre), 2) AS avg_gre,
-    ROUND(AVG(gre_v), 2) AS avg_gre_v,
-    ROUND(AVG(gre_aw), 2) AS avg_gre_aw
+    ROUND(AVG(gpa)::numeric, 2) AS avg_gpa,
+    ROUND(AVG(gre)::numeric, 2) AS avg_gre,
+    ROUND(AVG(gre_v)::numeric, 2) AS avg_gre_v,
+    ROUND(AVG(gre_aw)::numeric, 2) AS avg_gre_aw
 FROM applicants;
 """
 cur.execute(q3)
